[WIP] - Compartment files
