+++ example/App.js
@@ -1,43 +1,42 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- *
- * @format
- * @flow
- */
+// Dependencies
 import axios from 'axios';
 import React, { Component } from "react";
 import {
-  Platform,
-  StyleSheet,
+  createStackNavigator,
   Text,
   View,
   AppRegistry,
   Button
 } from "react-native";
+import Instructions from './components/instructions'
+import styles from './components/style'
 import * as RNEP from "@estimote/react-native-proximity";
-const instructions = Platform.select({
-  ios: "Press Cmd+R to reload,\n" + "Cmd+D or shake for dev menu",
-  android:
-    "Double tap R on your keyboard to reload,\n" +
-    "Shake or press menu button for dev menu"
-});
+
 
 // the IP address of the computer you are running server.js on with the PORT
-const localhost = "http://192.168.0.3:4000";
+const localhost = "http://192.168.1.70:4000";
 
-type Props = {};
-export default class App extends Component<Props> {
+export default class App extends Component {
+  // App State
+  state = {
+    testState: "State is Working",
+    connected: false,
+    tags:[],
+  };
+
+  // Create a new zone 
   zone2 = new RNEP.ProximityZone(1, "DigitalBillboard");
+
+  // Methods to interact with beacons
   enterAction = () => {
-    // zone2 = new RNEP.ProximityZone(10, "mint-leaf");
-    const ESTIMOTE_APP_ID = "digital-billboard-8gj";
-    const ESTIMOTE_APP_TOKEN = "9dff6faa96af5162dd8b20ec44e49ea5";
+    const ESTIMOTE_APP_ID = "billboards-o72";
+    const ESTIMOTE_APP_TOKEN = "563764b6a5ebe4d450cacda2f9438257";
     const credentials = new RNEP.CloudCredentials(
       ESTIMOTE_APP_ID,
       ESTIMOTE_APP_TOKEN
     );
 
+    // Notification
     const config = {
       notification: {
         title: "Exploration mode is on",
@@ -77,11 +76,12 @@ export default class App extends Component<Props> {
     };
   };
 
-  state = {
-    testState: "State is Working",
-    connected: false
-  };
+  createStackNavigator({
+    Home: { screen: Home },
+    Users: { screen: Users },
+  });
 
+  
   render() {
     return (
       <View style={styles.container}>
@@ -93,28 +93,9 @@ export default class App extends Component<Props> {
           {this.state.connected ? this.state.data.coupon : "No Deals Near"}
         </Text>
         <Text>{this.state.connected ? this.state.data.tag + ": tag nearby" : ""}</Text>
-        <Text style={styles.instructions}>{instructions}</Text>
+        <Text style={styles.instructions}>{Instructions}</Text>
         <Button onPress={this.enterAction} title="Get Started" />
       </View>
     );
   }
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "#F5FCFF"
-  },
-  welcome: {
-    fontSize: 20,
-    textAlign: "center",
-    margin: 10
-  },
-  instructions: {
-    textAlign: "center",
-    color: "#333333",
-    marginBottom: 5
-  }
-});
+};
