+++ example/components/screens/Home/home.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+import { View, Text, Button } from 'react-native';
+
+class Home extends Component {
+    render() {
+        return
+        <View>
+            <Text>This is the Home Screen</Text>
+        </View>
+    }
+};
+
+export default Home;
