+++ example/components/style/style.js
@@ -1,0 +1,25 @@
+import {
+    StyleSheet
+} from 'react-native'
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        justifyContent: "center",
+        alignItems: "center",
+        backgroundColor: "#AAC0AA"
+    },
+    welcome: {
+        flex: 2,
+        fontSize: 20,
+        textAlign: "center",
+        margin: 10
+    },
+    instructions: {
+        textAlign: "center",
+        color: "#333333",
+        marginBottom: 5
+    }
+});
+
+export default styles;
