+++ example/index.js
@@ -7,72 +7,6 @@ import { name as appName } from './app.json';
 AppRegistry.registerComponent(appName, () => App);
 
 // import * as RNEP from '@estimote/react-native-proximity';
-
-// // generate Estimote Cloud credentials for your app at:
-// // https://cloud.estimote.com/#/apps/add/your-own-app
-// const ESTIMOTE_APP_ID = 'digital-billboard-app-026';
-// const ESTIMOTE_APP_TOKEN = 'd6056fd23e22b958f7d478b2196e2c11';
-
-// // will trigger when the user is within ~ 5 m of any beacon with tag 'lobby'
-// // you can add tags to your beacons on https://cloud.estimote.com, in Beacon Settings
-// const zone1 = new RNEP.ProximityZone(5, 'mint-leaf');
-// zone1.onEnterAction = context => {
-// // context properties are:
-// // - attachments: all the key-value attachments assigned in Estimote Cloud to the beacon that triggered the action
-// // - tag: the tag used when defining the zone, repeated here for convenience
-// // - deviceIdentifier: Estimote-specific device identifier of the beacon that triggered the action
-//   console.log('zone1 onEnter', context);
-// };
-// zone1.onExitAction = context => {
-//   console.log('zone1 onExit', context);
-// };
-// zone1.onChangeAction = contexts => {
-// // onChange event gives you granular data about which exact beacons are in range
-// //
-// // imagine there are 2 beacons tagged 'lobby', to help cover the entire lobby area; here's an example sequence of events:
-// //
-// // 1. when you enter the range of the 1st one, you'll get:
-// // lobby onEnter
-// // lobby onChange with array [beacon1's context]
-// //
-// // 2. when you enter the range of the 2nd one, and are still in range of the 1st one:
-// // lobby onChange with array [beacon1's context, beacon2's context]
-// //
-// // 3. when you exit the range of the 1st one, but are still in range of the 2nd one:
-// // lobby onChange with array [beacon2's context]
-// //
-// // 4. when you finally exit the range of the 2nd one:
-// // lobby onChange with empty array []
-// // lobby onExit
-//   console.log('zone1 onChange', contexts);
-// };
-
-// // const zone2 = new RNEP.ProximityZone(5, 'conf-room');
-// // zone2.onEnterAction = context => {
-// //   console.log('zone2 onEnter', context);
-// // };
-// // zone2.onExitAction = context => {
-// //   console.log('zone2 onExit', context);
-// // };
-// // zone2.onChangeAction = contexts => {
-// //   console.log('zone2 onChange', contexts);
-// // };
-
-// // detecting proximity to Bluetooth beacons gives you information about the user's location, and so
-// // on both iOS and Android it's required to ask the user for permission to do that
-// //
-// // - on iOS, the user can choose between 'never', 'only when using the app' and 'always' (background)
-// //   - however, background support also requires that you enable the 'Uses Bluetooth LE accessories'
-// //     Background Mode for your app
-// //   - you can do that in Xcode project settings, on the Capabilities tab
-// //   - you might also need to explain/defend your app's background usage during the App Store review
-// //
-// // - on Android, it'll be a simple 'yes/no' popup, which is equivalent to 'never' and 'always'
-// //   - however, to have it work in the background, you're also required to show a notification, so
-// //     that the user knows that the app keeps running/detecting beacons even if they close it
-// //   - see the `config` section below for how to enable/configure such notification
-// //
-// // see also: 'Location permission' and 'Background support' sections in the README
 // RNEP.locationPermission.request().then(
 //   permission => {
 //     // `permission` will be one of RNEP.locationPermission.DENIED, .ALWAYS, or .WHEN_IN_USE
